requirements.txt
# Ana gereksinimler
flask>=2.0.1
python-dotenv>=0.19.0
Pillow>=8.4.0
opencv-python>=4.5.5

# TensorFlow Ekosistemi
numpy>=1.24.3
tensorflow>=2.13.0
tensorflow-hub>=0.14.0
keras>=2.13.0
keras-ocr>=0.9.1

# BERT ve NLP
transformers[tf]>=4.21.0

# Utility
python-magic-bin>=0.4.14
tqdm>=4.62.3
requests>=2.31.0
urllib3<2.0.0

# Performance
gunicorn>=20.1.0
uvicorn>=0.15.0